# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AddProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  dto.AddTask:
    properties:
      assigneeid:
        type: integer
      description:
        type: string
      due_date:
        type: string
      parent_task_id:
        type: integer
      project_id:
        type: integer
      titel:
        type: string
    required:
    - assigneeid
    - description
    - due_date
    - project_id
    - titel
    type: object
...
  dto.EditTask:
    properties:
      description:
        type: string
      id:
        type: integer
      priority:
        enum:
        - Low
        - Medium
        - High
        - Hihg
        type: string
      status:
        enum:
        - Todo
        - In
        - Progress
        - Done
        type: string
      titel:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
  dto.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RefreshTokenDto:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  dto.UpdateProject:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
...
  dto.UserDto:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: mohammed.qassem@baly.iq
    name: Mohammed
  description: Task Management System API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Task Management API
  version: "1.0"
paths:
  /Project/{id}/Delete/{projectid}:
    delete:
      consumes:
      - application/json
      description: Delete the projects in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: project ID
        in: path
        name: projectid
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete the projects in the workspace
      tags:
      - Project
  /Project/{id}/Get:
    get:
      consumes:
      - application/json
      description: get projects in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: get projects in the workspace
      tags:
      - Project
  /Project/{id}/Update:
    put:
      consumes:
      - application/json
      description: update the project in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateProject Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProject'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: update the project in the workspace
      tags:
      - Project
  /Task/{id}/Create:
    post:
      consumes:
      - application/json
      description: Create a new task in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: AddTask Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddTask'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new task in the workspace
      tags:
      - Task
  /Task/{id}/Delete/{task_id}:
    delete:
      consumes:
      - application/json
      description: Delete the task in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete the task in the workspace
      tags:
      - Task
  /Task/{id}/Edit:
    put:
      consumes:
      - application/json
      description: Edit the task in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: EditTask Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EditTask'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Edit the task in the workspace
      tags:
      - Task
  /Task/{id}/GetAll:
    get:
      consumes:
      - application/json
      description: Get all the task in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task Status
        in: query
        name: Status
        type: string
      - description: ProjectId Number
        in: query
        name: ProjectId
        type: integer
      - description: Task Priorty
        in: query
        name: Priorty
        type: string
      - description: AssigneeId
        in: query
        name: AssigneeId
        type: integer
      - description: Serch
        in: query
        name: Serch
        type: string
      - description: SortBy
        in: query
        name: SortBy
        type: string
      - description: Order
        in: query
        name: Order
        type: string
      - description: Limit Number
        in: query
        name: Limit
        type: integer
      - description: Offset Number
        in: query
        name: Offset
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all the task in the workspace
      tags:
      - Task
  /login:
    post:
      consumes:
      - application/json
      description: Sign In to youer account
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses: {}
      summary: Login
      tags:
      - Authentication
  /project/{id}:
    get:
      consumes:
      - application/json
      description: Get a project by its ID
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: projectid
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a project by ID
      tags:
      - Project
  /project/{id}/create:
    post:
      consumes:
      - application/json
      description: Create a new project in the workspace
      parameters:
      - description: WorkSpace ID
        in: path
        name: id
        required: true
        type: integer
      - description: AddProjectRequest Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddProjectRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new project in the workspace
      tags:
      - Project
  /refreshToken:
    post:
      consumes:
      - application/json
      description: refresh youer jwt
      parameters:
      - description: RefreshToken request
        in: body
        name: requst
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDto'
      produces:
      - application/json
      responses: {}
      summary: RefreshToken
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserDto'
      produces:
      - application/json
      responses: {}
      summary: Register new user
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"